hooks: tensorflow: rework the hook to improve compatibility across different versions
Clean and up and rework the tensorflow hook to improve compatibility
across different tensorflow versions.

Currently, the `tensorflow` hook collects the submodules from
`tensorflow_core` in a rather roundabout way: by specifying
`tensorflow_core` as a hidden import, and relying on this to pull in
`tensorflow_core.python`. The hook for `tensorflow_core.python` is
the one that actually collects data and submodules from `tensorflow_core`.

This mechanism is brittle, and seems to break on:
* tensorflow 1.15.x: due to import rerouting between `tensorflow_core`
and `tensorflow`, the `tensorflow_core.python` is not pulled in
* tensorflow 2.2.0+: the `tensorflow_core` module has been removed, and
its contents now reside within `tensorflow` itself.

To fix this issue and improve compatibility across different versions
of tensorflow, the `tensorflow_core.python` hook is removed, and
`tensorflow` hook is modified to collect data and submodules from
either `tensorflow` or `tensorflow_core`, depending on version:
* 1.14.x and earlier: collect from `tensorflow`
* 1.15.x - 2.1.x: collect from `tensorflow_core`
* 2.2x and later: collect from `tensorflow`