Rework the tensorflow hook to improve compatibility across different tensorflow versions (#46)
* hooks: tensorflow: Fix the hook to be compatible with versions 1.15.0 to 2.3.0.

Clean and up and rework the tensorflow hook to improve compatibility
across different tensorflow versions.

Currently, the `tensorflow` hook collects the submodules from
`tensorflow_core` in a rather roundabout way: by specifying
`tensorflow_core` as a hidden import, and relying on this to pull in
`tensorflow_core.python`. The hook for `tensorflow_core.python` is
the one that actually collects data and submodules from `tensorflow_core`.

This mechanism is brittle, and seems to break on:
* tensorflow 1.15.x: due to import rerouting between `tensorflow_core`
and `tensorflow`, the `tensorflow_core.python` is not pulled in
* tensorflow 2.2.0+: the `tensorflow_core` module has been removed, and
its contents now reside within `tensorflow` itself.

To fix this issue and improve compatibility across different versions
of tensorflow, the `tensorflow_core.python` hook is removed, and
`tensorflow` hook is modified to collect data and submodules from
either `tensorflow` or `tensorflow_core`, depending on version:
* 1.14.x and earlier: collect from `tensorflow`
* 1.15.x - 2.1.x: collect from `tensorflow_core`
* 2.2x and later: collect from `tensorflow`

* tests: added two new tests for tensorflow

Added two new tests for tensorflow in order to verify that at least
the basic tensorflow functionality is both importable and functional:
* a basic use of a Conv2D layer
* a basic model training and validation on MNIST dataset

* tests: perform tensorflow tests only in onedir mode

Skip tensorflow tests in modes other than onedir mode. With recent
tensorflow versions, the resulting dist size could be up to 1.5 GB,
which may cause onefile tests to timeout due to relatively long
compression and decompression time.

* hooks: tensorflow: minimize the use of `is_module_satisfies()`

A minor cleanup; use `is_module_satisfies()` to define version
related booleans, and then use those in the if/else blocks to make
them easier to read.

* tests: moved the two complex tensorflow tests to external script files

Also grouped the tensorflow tests together in the test_libraries.py