hooks: tensorflow: prevent duplication of _pywrap_tensorflow_internal
On some tensorflow versions (2.3.0 thus far), the
_pywrap_tensorflow_internal extension module ends up duplicated.
We pick it up as both an extension module (placed in
`tensorflow/python` directory) and as a shared library (placed in
program's root directory). This increases the program size, and
also prevents the program from running under macOS.

In the problematic versions, the extension module seems to be
picked up as `tensorflow.python._pywrap_tensorflow_internal`,
while in others it is picked up as just ˙_pywrap_tensorflow_internal`.
So we try to work around the duplication by removing the problematic
entry from `hiddenimports` and adding it to `excludedimports`.

Fixes #49.