Create new CI workflow OneShot test.
This action is for one-off bulk testing against multiple versions
of a single package. With it, hook authors can fully test the hook
they are working on by invoking this action and providing a range
of library requirements to install.

Invoking this action is only done manually via:

  actions -> workflows -> Oneshot test -> Run workflow
  -> Fill in the options it asks for -> Run workflow

I may create a script that triggers the above using the Github API
if we find the *fill in the options* step gets a bit tedious.