Add hooks for skimage.feature and skimage.graph (#52)
A battery of tests that individually import each of skimage submodules. The list of tests is based on submodules available
in the latest releases (0.16.x and 0.17.x).

Add hook for skimage.feature; it's needed for hidden import of skimage.feature._orb_descriptor_positions.

Add hook for skimage.graph; it's needed for hidden import of skimage.graph.heap.