tests: add basic import test for scikit-learn
A battery of tests that individually import each of sklearn
submodules. The list of tests is based on submodules available
in the latest release (0.23.x), but it has been unchanged since
0.21.

Incidentally, scikit-learn 0.21 supports Python 3.5-3.7, scikit-learn
0.22 supports Python 3.5-3.8, while the latest versions require
Python 3.6 or newer.

Therefore, the test supports scikit-learn >= 0.21, and the same
versions will be targeted by our hooks.