hooks: pydantic: add missing dataclasses and distutils.version hidden imports (#81)
This standard library module is imported in some functions from
pydantic.dataclasses module, so the module-not-found-error occurs
only at the time of function call. Fixes pyinstaller/pyinstaller#5432.

Add distutils.version as a hidden import, as older versions of pydantic (prior to 1.4) import distutils.version
inside pydantic.version, so add it to the list of hidden imports.
This fixes the test failures for pydantic from 1.0 to 1.3 on linux
(on other platforms, presumably the distutils.version is also pulled
in by some other dependency).