pygraphviz hook: add Linux support
See
<https://github.com/pyinstaller/pyinstaller-hooks-contrib/pull/86#pullrequestreview-584684190>,
this was not done before due to an implicit list of executables, which
is no longer a problem.

Keep the logic as "if windows, else if macos, else", so it can cover not
only Linux, but other similar cases (BSD, etc).

Don't use findLibrary() on macOS, where it is not available and not
needed to figure out the distro-specific lib vs lib64 difference.