CI/CD: Oneshot: Implicitly wrap each requirement in double quotes. (#136)
When given a requirement like `humanize>3` bash will interpret the `>`
as the *stream to file* operator so that `pip install humanize>3`
will really run `pip install humanize` and pipe the output to a file
called `3`. Wrapping with quotes (`pip install "humanize>3"`) fixes this
but it complicates the CI launching UI because, unless you know why the
quotes are needed, you're guessing where/when to use them.

Now each requirement given is internally quoted automatically so that
there is never any need to manually quote parameters. Input syntax now
mostly mirrors that of PEP 508 requirements files with the differences
being that `,` splits jobs and therefore can't be used anywhere else
and that single requirements like 'requests >= 2' must be written
without spaces.