tests: fix distribution version checks scikit-image and scikit-learn
Call is_module_satisfies with distribution name (scikit_learn)
instead of package name (sklearn), because running `pip
install sklearn` actually installs `sklearn==0.0` and pulls
in latest version of `scikit-learn`. So if scikit-learn is
accidentally installed this way, the tests will be skipped
because the version checks based on package name will fail.

Make the same change for scikit-image (scikit_image vs
skimage) for the sake of consistency (although in this
case, installing `skimage` results in an error telling
user to install `scikit-image` instead).