hooks: mariadb: always collect decimal module as hidden import
Always collect the `decimal` module as a hidden import, even
if it seems to be automatically picked up on recent python versions.
That likely happens due to import in some other module (e.g., from
standard library), but may be subject to change in the future,
so we should not be relying on it. Besides, the hidden import does
no harm, even if it is redundant.