hooks: cv2: collect Qt fonts data and plugins from linux PyPI wheels
In linux PyPI opencv-python wheels, the cv2 extension is linked
against Qt, and the wheel bundles a basic subset of Qt shared
libraries, plugins, and font files. This is not the case on other
OSes (presumably native UI APIs are used by the OpenCV HighGUI
module), nor in the headless PyPI wheels (opencv-python-headless).

The bundled Qt shared libraries should be picked up automatically
due to binary dependency analysis, but we need to collect plugins
and font files from the qt subdirectory.