hooks: update pywintypes and pythoncom hooks
Have pywintypes and pythoncom hook collect their corresponding
DLLs into pywin32_system32 sub-directory if the DLLs originally
came from site-packages/pywin32_system32 directory (which is the
case with PyPI wheel, but not with Anaconda package). This
prevents duplication of pywintypesXY.dll when the latter is
also collected as a part of link-time binary dependency analysis,
which starting with PyInstaller 5.4 attempts to preserve the parent
directory structure.