hooks: prevent pyqtgraph hook from recursing into pyqtgraph.examples
Prevent `pyqtgraph` hook from recursing into `pyqtgraph.examples`
while scanning for submodules. While the obtained list of submodules
is subsequently filtered to include only templates, import of
`pyqtgraph.examples` during scan triggers instantiation of
`QtWidgets.QApplication`. This, in turn, requires a valid X/Wayland
session on linux and and lack thereof results in an error that
crashes the python subprocess in which we scan for submodules.