Update hook-tensorflow.py
In the recent TF 2.12 release, there is a new file `libtensorflow_cc.so.2` which now appears twice in the final output: once correctly in the root directory (as 'binary') and once incorrectly in the `tensorflow` subdirectory (as 'data' file). This new file is where they moved the majority of the code, it's over 500 MB, so this duplication causes a lot of overhead.

To reproduce the issue, install `tensorflow==2.12` or `tensorflow-cpu==2.12` and package a file with only `import tensorflow`.
With these changes the file only appears once in the final output and the package still works as expected.

I have tested this change on Linux only.