hook for skimage
starting skimage 0.20, I believe, many things are lazy-loaded.
To get my project with skimage to work with pyinstaller, I had to upgrade lazy_loader -> 0.2 and include this hook file.
hook file created by reading https://github.com/scikit-image/scikit-image/issues/6772#issuecomment-1449671566 and guessing
Tests did not run for me. There is already a test_skimage in the suite, but `pytest -k 'test_skimage'` would either deselect all tests or (if I removed some decorators in front of test_skimage) run the function with a bunch of parameters that are not listed in the parametrize mark.