tests: test great-expectations only on python < 3.11
great-expectations does not officially support python 3.11, yet,
and while earlier versions did not seem to have problems with it,
the 0.16.10 made some changes to pyspark compatibility wrapper
that seem to be incompatible with python 3.11. It raises an error
when pyspark is unavailable, even though it is an optional dependency,
and works just fine without it on earlier python versions.