hooks: Fix googleapiclient.model missing docs and improve test (#596)
The hook googleapiclient.model was raising this warning when running
Pyinstaller: WARNING: collect_data_files - skipping data collection for module
'googleapiclient.discovery' as it is not a package. Needing workarounds to run
the api client, see googleapis/google-api-python-client#876. But this can be
fixed by pointing it to the correct package googleapiclient.discovery_cache.

The test also received an improvement. The current test only run from
googleapiclient.discovery import build which is not sufficient because the error
happens when calling the build function. Calling this function doesn't require a
valid API key.