hooks: torch: add support for external nvidia.* packages on linux
The contemporary PyPI torch wheels for linux use CUDA libraries
that are installed via `nvidia-*` packages. Therefore, attempt to
convert the `nvidia-*` requirements from the `torch` metadata
into hidden imports. This way, we can provide hooks for `nvidia.*`
packages that collect the shared libs, in case any of them are
dynamically loaded (which currently seems to be the case with
some of the shared libraries from `nvidia.cudnn`).