tests: add test for torchtext that uses torchscript
Add a test for torchext that uses a tokenization transform from
Berta Encoder. The test shows the need to collect binaries from the
torchtext package, and, due to use of torchscript, also shows the
need to collect source .py files.

We perform only tokenization part of processing, in order to avoid
having to download the whole model (~240 MB).