hooks: add a hook for fastai
Add a hook for fastai, which ensures that the package's source .py
files are collected, as they are required for TorchScript.

Add a test based on the "building models from tabular data" example
from https://docs.fast.ai/quick_start.html, which demonstrates
the need for the source .py files.