hooks: add hook for lightning
Add hook for (PyTorch) `lightning`. Currently, the main functionality
is to ensure that the `version.info` file from the package is
collected.

We do not collect source .py files, as it seems that even if
`lightning.LightningModule.to_torchscript()` is used, it requires
the source where the model inheriting from `lightning.LightningModule`
is defined, rather than `lightning`'s own sources. We can always
add source .py files collection later, if it proves to be necessary.