hooks: tensorflow: rework the tensorflow version check
Determine the tensorflow's dist name using list of potential
candidates, and if a dist is found, retrieve the version from
it. Otherwise, fall back to reading `tensorflow.__version__`.

The `tensorflow` package is available in several variants, and
sometimes the `tensorflow` dist installs a separate dist (e.g.,
`tensorflow-intel` on Windows); but the user can install this
separate dist without installing the "top-level" `tensorflow`
one. In PyInstaller v5 and earlier, the `is_module_satisfies`
fell back to querying `tensorflow.__version__` if the dist could
not be found - in v6, the implementation of `is_module_satisfies`
(or rather, `check_requirement`) checks only the metadata.

As an added bonus, the direct version comparisons are nicer to
read than the comparisons against `tf_pre_1_15_0`, `tf_post_1_15_0`,
etc.