hooks: remove hook for google.api
Remove hook for `google.api`, which erroneously assumes that presence
of the `google.api` namespace package implies availability of the
`google-api-core` dist.

This results in an error when, for example, only
`googleapis-common-protos` is installed in the build environment.

The collection of `google-api-core` metadata should be sufficiently
handled by the hook for `google.api_core`.