tests: force stdout/stderr encoding in tensorflow mnist test to utf-8
The latest version of `tensorflow` (2.16) and `keras` (3.0) upgraded
the interactive progress bar (displayed during dataset download,
during model fitting, etc.) with fancier version that uses Unicode
arrows (←). For this to work, `sys.stdout` must be using utf-8
encoding.

On Windows, this is default when stdout/stderr are attached to
console, but it is not the case when they are redirected to pipes,
as in that case, python defaults to system locale encoding.

PyInstaller's `pyi_builder` test fixture runs the built executable
via `subprocess.Popen` with stdout and stderr redirected to pipes,
so stdout and stderr end up using system locale encoding (for
example, cp1252), which results in the following error:

```
[960] LOADER: Running pyi_lib_tensorflow_mnist.py
Traceback (most recent call last):
  File "pyi_lib_tensorflow_mnist.py", line 27, in <module>
  File "keras\src\datasets\mnist.py", line 60, in load_data
  File "keras\src\utils\file_utils.py", line 291, in get_file
  File "urllib\request.py", line 267, in urlretrieve
  File "keras\src\utils\file_utils.py", line 283, in __call__
  File "keras\src\utils\progbar.py", line 182, in update
  File "keras\src\utils\io_utils.py", line 98, in print_msg
  File "encodings\cp1252.py", line 19, in encode
  UnicodeEncodeError: 'charmap' codec can't encode characters in position 33-52: character maps to <undefined>
[960] Failed to execute script 'pyi_lib_tensorflow_mnist' due to unhandled exception!
```

Running the same executable directly in command prompt succeeds,
as in that scenario, stdout and stderr end up using utf-8.

Therefore, as a work around, have the test application script check
the encoding of stdout/stderr, and reconfigure it, if necessary.