hooks: numba: ensure numba.cloudpickle.cloudpickle_fast is collected
`numba` 0.59.0 updated its vendored version of `cloudpickle` to 3.0.0;
this version keeps `cloudpickle_fast` module around for backward
compatibility with existing pickled data, but does not import it
directly anymore.

Therefore, we need to add it to hidden imports to ensure that it is
collected and available when unpickling data that was pickled with
older version of `numba`.