hooks: cryptography: collect OpenSSL >= 3.0.0 modules, if available
Have the `cryptography` hook collect OpenSSL >= 3.0.0 modules from
`ossl-modules` directory, if available. Add a run-time hook that
sets `OPENSSL_MODULES` environment variable to the bundled
`ossl-modules` directory.

This is required for `cryptography` to work with recent OpenSSL
versions, where legacy algorithms were put into separate legacy
module. This module can be a separate shared library (e.g., on
Linux in general, on Windows in msys2/MinGW, and on macOS in
Homebrew), so we need to make sure we collect it.