ci: use macos-12 and ubuntu-22.04 instead of -latest
Use `macos-12 runner`, to run on x64_64 Monterey instead of arm64
Sonoma, which macos-latest now points to. This fixes issues with
packages that are not yet providing arm64 wheels (e.g., `adbutils`)
or packages that do not provide arm64 wheels for older python
versions (e.g., `lxml` for python 3.8). In addition, it enforces
consistency between this CI pipeline and the one used by the
main repository.

Similarly, for the sake of consistency with the main repository,
use `ubuntu-22.04` runner instead of `ubuntu-latest` one.