hooks: tensorflow: suppress binary dependency analysis symlinks
On Linux, use the new mechanism for suppressing creation of symbolic
links to top-level application directory for binaries discovered during
binary dependency analysis, for the following `tensorflow` shared
libraries: `libtensorflow_cc.so.2`, `libtensorflow_framework.so.2`,
and `_pywrap_tensorflow_internal.so`.

Having symbolic links to these libraries in the top-level application
directory seems to confuse tensorflow about its location (likely
because code in one of the libraries looks up the library file's
location, but does not fully resolve it), which in turn prevents it
from finding the collected CUDA libraries in the nvidia/cu* package
directories.

This fixes CUDA not being found by tensorflow in frozen Linux builds.
It requires PyInstaller >= 6.11, which added the symlink suppression
mechanism; on earlier versions, the corresponding part of the hook
code is no-op.