hooks: nvidia.*: suppress generation of symlinks for shared libs
On Linux, have the hooks for `nvidia.*` pacakges use the
`bindepend_symlink_suppression` hook attribute to prevent
PyInstaller's binary dependency analysis from creating symlinks
to top-level application directory for shared libraries discovered
in these packages during binary dependency analysis.

Some of the shared libraries from these packages are loaded
dynamically (e.g., `dlopen` at run time), and they are sought in
locations that are relative to the location of the "main" shared
library that is performing the loading and search (and which is
linked to python extensions). However, the location of these
shared libraries is not fully resolved, so a symlink causes
the library to misidentify its location and search for other
shared libraries in the wrong locations.

For example, when using `ultralytics` package, the symlink for
`libcudnn_graph.so.9` from `nvidia/cudnn/lib` directory prevents
discovery of the `libcudnn_engines_precompiled.so.9` in the same
directory (because it is sought in top-level application directory
instead of its real location). This results in the following
run-time error: `RuntimeError: CUDNN_BACKEND_TENSOR_DESCRIPTOR
cudnnFinalize failed cudnn_status: CUDNN_STATUS_NOT_INITIALIZED`.

Therefore, prevent creation of symlinks to top-level application
directory for all `nvidia.*` packages. In unfrozen python, these
are either preloaded via dlopen() from pre-determined location, or
linked against with relative RPATH; so there is not need for symlinks
to the top-level application directory in the first place.