hooks: pydicom: do not collect data files from `test_files`
Avoid collecting all data files from `test_files` directory, even
the ones that are referenced `_DATASETS` dictionary that is
constructed during initialization of `pydicom` module in v3.0.0.

It seems that these files are not listed in `urls.json` so even if
they are missing, not attempt is made to download them. If anyone
requires these test files in their frozen application, they should
explicitly collect the files themselves (same as with earlier
versions of `pydicom`).