hooks: toga_winforms: collect only relevant runtime part
Collect only the `win-x86`/`win-x64`/`win-arm64` sub-directory of
`toga_winforms\libs\WebView2\runtimes` that is relevant to the
target architecture (i.e., the architecture of python under which
PyInstaller is running).

Consolidate collection of DLLs and data files to use a single
`collect_data_files` call. While strictly speaking, we should be
assinging DLLs to `binaries` instead of `datas` (and collect them
using `collect_dynamic_libs`), in the case of these DLLs, it does
not matter much. And especially not with PyInstaller 6.x, which
performs automatic binary-vs-data classification.