hooks: toga: move backend metadata collection into backend hooks
Remove collection of entry-points in the `toga` hook. The
`setuptools_scm.parse_scm` should be irrelevant to the frozen
application. The `toga.backends` entry point should be avoided
because technically, we are aiming to collect exactly one backend
instead of all available ones (in the unlikely event that more
than one is installed); instead, the metadata for the backend
dist should be collected by the backend hook.