hooks: tensorflow: automatically raise recursion limit
Several tests on the CI are throwing recursion error with
`tensorflow` 2.18.0 installed, so have `tensorflow` hook automatically
raise the recursion limit to 5000 (unless it is already at higher
value).

The problem does not seem to be tied to this `tensorflow` version
(building in a clean environment with just `tensorflow` and its
dependencies installed seems to pass without errors), but rather
to specific import chains that involve `tensorflow` but also depend
on the build environment (i.e., other packages installed in it).