tests: update test_apscheduler
Remove `pytz` dependency; `apscheduler` 3.11.0 does not install
`pytz˙ as its dependency anymore. But the only place it is used
in test is for `pytz.UTC`, which we can easily replace with
`datetime.timezone.utc`.

The current test program does not run under `apscheduler` 3.11.0
anymore (even when unfrozen), as it raises `RuntimeError: no running
event loop`. Rework the test code to match the official asyncio example:
https://github.com/agronholm/apscheduler/blob/3.x/examples/schedulers/asyncio_.py