hooks: rewrite pygraphviz hook
Rewrite `pygraphviz` hook to fix discovery and collection of `graphviz`
files under various Linux distributions, in Anaconda environments
(Windows, Linux, and macOS), and msys2 environments (Windows).

We now discover active `graphviz` installation on all OSes by looking
for `dot` program in `PATH`. To find the location of shared libraries
and the plugin directory on non-Windows, we now perform binary
dependency analysis on the `dot` executable and look for known
`graphviz` shared libraries in the results.

This improves the discovery on various platforms, and also fixes the
issue where system-installed (on Linux) or Homebrew-installed
(on macOS) `graphviz` shared libraries and plugins would end up being
collected  instead of Anaconda-packaged ones when using Anaconda
environment with Anaconda-packaged `graphviz`.