Revert "Update eccodes from 2.39.2 to 2.40.0"
This reverts commit d72954231efea9e426965d3cdf8813fda1bb2efd.

`eccodes` depends on `findlibs`, whose v0.1.0 release is incompatible
with python < 3.10 despite not being marked as such. In practice,
`findlibs` is not used (not even imported) when `eccodes` shared
libraries are bundled with binary wheel, but the issue is triggered by
our run-time hook for `findlibs`, which imports the package.

Will deal with this in a separate PR, together with enabling the
installation of `eccodes` on Windows (looks like binary wheels are
available since v2.39.0).