ci: update linux runner for compatibility with for PyGObject >= 3.52.1
Some of the tested packages (e.g., `toga` / `toga-gtk`) end up
pulling `PyGObject`, which recently (v3.52.1) switched to
`girepository-2.0` (i.e., requires `GIRepository` typelib 3.0
instead of 2.0).

Therefore, we need to switch the runner to `ubuntu-24.04` (which
we should do anyway, to match similar switch made in the main
repository), and ensure that `libgirepository-2.0-dev` and
`gir1.2-girepository-3.0` packages are installed via `apt`
(as well as `libcairo2-dev`, which seems to be required for
building and installing `pycairo`, a dependency of `PyGObject`).