hooks: update hook for travertino to explicitly collect package's metadata
Update hook for `travertino` to explicitly collect the package's
metadata in order to fix compatibility with `travertino` v0.5.0.

With earlier `travertino` releases, we were able to detect the
attempted use of metadata and collect it automatically; with
v0.5.0, the corresponding codepath was refactored to no longer
call `importlib.metadata.version()` with const parameter (but
rather with variable), which precludes auto-detection on our
part.