hooks: torch: fix path matching in MKL DLL collection part of the hook
Fix path matching for MKL DLLs in `torch` hook to work with either
POSIX-style or Windows-style separators, as either can appear in the
metadata's `RECORD` entries.

The `importlib_metadata.PackagePath.match()` does not seem to properly
normalize the separator, so we now first resolve the file's location
(which yields a `pathlib.Path`) and perform matching on resolved path.