hooks: add hooks for submodules of vtkmodules from vtk dist
Add hooks for all submodules of the `vtkmodules` package, which
is installed by the `vtk` dist.

Unfortunately, the `.pyi` files that accompany the extension modules
do not provide all dependencies, so we need to maintain per-module
list of hidden imports / dependencies ourselves.

Add individual import tests for all said submodules, so that we
will know when the said list of dependencies need to be updated.