ci: download and install the TLB file for PyTaskbar
PyTaskbar 0.1.0 does not ship TaskBarLib.tlb with its wheels anymore
(in contrast to earlier releases under the PyTaskbarProgress name);
apparently users are supposed to download it separately and copy it
to their project's directory. Copying to the package's directory also
seems to work, and is what our hook is expecting. So until we see
what directions subsequent releases take, download and install
the file in a CI pipeline step.