test requirements: remove entries with "pyup: ignore" comment
These entries should never be updated; since our CI is based
on the diff of the requirements file, it makes makes no sense
to keep them around.

This is in contrast to the main repository's requirements file,
where we do keep pinned older package versions for older python
versions as well (but we always install whole set of packages, not
just the diff).