Adjust the automatically-raised recursion limit
The recursion limit test in PyInstaller main repository seems to push
the recursion level to the limit; and if the limit is set to 5000, this
crashes python 3.8 - 3.10 on Windows and 3.9 that is (at the time of
writing) available under Cygwin.

Further investigation revealed that Windows builds of python 3.8 and 3.10
handle recursion up to level ~2075, while the practical limit for 3.9 is
between 1950 and 1975. Therefore, for affected combinations of platforms
and python versions, use a conservative limit of 1900 - if only to avoid
issues with the recursion limit test in the main PyInstaller repository...

For other platforms and python versions, leave the default limit at
the 5000 as per earlier commit, although it is likely that it could
be lowered as well.