rthooks: findlibs: avoid overriding `pkg_name` argument
Our override for `findlibs.find()` makes no use of the `pkg_name`
argument and should avoid overriding the `None` value; instead, the
argument should be passed to the original implementation as is.

This fixes the behavior with `findlibs` >= 0.1.0; while the older
versions (e.g., 0.0.5) used `lib_name` as the default value for
`pkg_name`, 0.1.0 and later use `lib_name` + `'lib'`.

In addition, future-proof our override by having it use `*args`
and `**kwargs` to pass all additional arguments to the original
implementation.