tests: bump eccodes version to 2.43.0
Bump the tested version of `eccodes` to 2.43.0; this version does
not work with python 3.9 anymore, because on non-Windows, shared
libraries are bundled by external `eccodeslib` package, and
`findlibs` is imported to discover them. The `findlibs` package,
in turn, uses typing annotations that implicitly require
python >= 3.10, and the package's metadata fails to declare this
requirement.