hooks: harden blib2to3 and torch hooks against missing RECORD in metadata
Harden `blib2to3` and `torch` hooks against missing `RECORD` in
distribution metadata; this would cause `Distribution.files()`
to return a `None`, which is non-iterable. So we need to use
`Distribution.files() or []` pattern to guard against that.

For example, very few of Fedora python packages seem to provide
`RECORD` entry, and the `blib2to3` hook was observed to run into
non-iterable error with Fedora-packaged `black`.