tests: make test_apscheduler invariant to the TZ environment variable
Make `test_apscheduler` invariant to the `TZ` environment variable,
by specifying the timezone when constructing `AsyncIOScheduler`.
For the purposes of the test, the timezone itself should not matter,
so use UTC.

The `TZ` environment may contain values such as `UTC0` or `GMT+12`,
which is valid but not supported by `tzlocal` that is used by
`apscheduler` behind the scenes.