hooks: update hooks and test for langchain >= 1.0.0
In `langchain` v1.0.0, the API has changed completely.

The old interfaces are available in optional `langchain_classic`
package, for which we need to collect data files (same as we had to
for `langchain` with versions prior to 1.0.0).

Put the old test under version check, and create a counterpart for
`langchain_classic`.

In addition, try to set up a basic test for the new `langchain` itself;
since this requires access to an LLM model, have the test use `ollama`,
and exit if it cannot connect to the local instance. This way, we can
at least check on the CI that everything works up until the point
where access to the model is required (and locally, it is possible
to run complete test by running a docker container with the model).