hooks: add hook for pymeshlab
Add hook for `pymeshlab` to ensure that wheel-provided Qt plugins
are collected in addition to Qt libraries (which are resolved by
binary dependency analysis), and that meshlab plugins are collected.

Add basic test that tries to import the package, generate a mesh,
and save it to a .ply file to ensure that plugins are working.